requirements.txt
asgiref==3.5.2
autopep8==1.7.0
bleach==5.0.1
Django==4.1.2
django-cleanup==6.0.0
django-summernote==0.8.20.0
Pillow==9.2.0
pycodestyle==2.9.1
python-decouple==3.6
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
tzdata==2022.5
webencodings==0.5.1
